[Game] Fixed SCScheduleItemUpdatePacket: - Divine Clock now works; - it is not yet possible to receive gifts online in the game;

+++ AAEmu.Game/Core/Packets/G2C/SCScheduleItemUpdatePacket.cs
@@ -16,7 +16,7 @@ public class SCScheduleItemUpdatePacket : GamePacket
 
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(_scheduleItems.Count);
+        stream.Write((byte)_scheduleItems.Count);
         foreach (var item in _scheduleItems)
         {
             stream.Write(item);
